YouTube downloads via yt-dlp: video/audio + playlist with per-item progress
Add a "YouTube" menu (video with audio / audio only / video only /
playlist-or-channel) backed by yt-dlp + ffmpeg. Downloads appear as normal
list rows and share the existing Stop / Stop-All cancellation path.

- youtube.rs: single video/audio runner. Probes title+size up front (IDM-style
  dialog), spawns yt-dlp, streams progress into the engine as a normal row.
- youtube_playlist.rs: playlist/channel manager + progress window. Each entry
  downloads as its own yt-dlp subprocess (sequential) so individual items can be
  Resumed/Deleted. The progress window's "Progress" column is custom-drawn
  (pl_customdraw / NM_CUSTOMDRAW) to show a per-item bar + percent for the item
  currently Downloading; an aggregate bar at the bottom tracks done/total.
  Dark-mode aware track/fill/text colors.
- engine.rs: primitives for externally-run downloads (alloc_id / emit /
  register / unregister) so yt-dlp jobs reuse the same cancel tokens as native
  downloads (Stop / Stop All / stop_queue).
- options.rs + settings.rs: yt-dlp path / ffmpeg path fields (empty = auto:
  app folder -> PATH), each with a Browse... button.
- tasks.rs: pick_exe() file picker for the two Browse buttons.
- store.rs: set_name() to update the aggregate playlist row label.
- gui.rs: YouTube popup menu, command handlers, playlist-row routing, and a
  disabled/tooltip state when yt-dlp is missing.
- dark.rs: add # Safety docs to unsafe pub fns; minor dwItemSpec type cleanup.

Compiles clean (cargo check, no warnings). Not yet visually verified in a
release build.

+++ crates/adm-app/src/dark.rs
@@ -46,6 +46,9 @@ unsafe fn cached_brush(slot: &Mutex<isize>, color: (u8, u8, u8)) -> HBRUSH {
 
 /// Terapkan dark ke window dialog + semua anaknya. Panggil SETELAH semua kontrol
 /// anak dibuat (mis. tepat sebelum `ShowWindow`). No-op bila tema terang.
+///
+/// # Safety
+/// `hwnd` harus window valid; panggil dari thread pemilik window (UI thread).
 pub unsafe fn apply(hwnd: HWND) {
     if !is_dark() {
         return;
@@ -85,6 +88,9 @@ unsafe extern "system" fn theme_child(child: HWND, _: LPARAM) -> BOOL {
 /// Tangani `WM_CTLCOLOR*` (STATIC/EDIT/BTN/LISTBOX/DLG/SCROLLBAR). Kembalikan
 /// `Some(brush)` sbg LRESULT bila dark aktif; `None` → biarkan proc pakai default.
 /// `wparam` = HDC kontrol.
+///
+/// # Safety
+/// `wparam.0` harus HDC valid (dari pesan `WM_CTLCOLOR*` yang sedang ditangani).
 pub unsafe fn ctlcolor(msg: u32, wparam: WPARAM) -> Option<LRESULT> {
     if !is_dark() {
         return None;
@@ -106,6 +112,9 @@ pub unsafe fn ctlcolor(msg: u32, wparam: WPARAM) -> Option<LRESULT> {
 
 /// Isi latar dialog dengan warna gelap saat `WM_ERASEBKGND`. Kembalikan
 /// `Some(LRESULT(1))` bila ditangani (dark), `None` bila terang.
+///
+/// # Safety
+/// `hwnd` harus window valid dan `wparam.0` HDC valid (dari `WM_ERASEBKGND`).
 pub unsafe fn erasebkgnd(hwnd: HWND, wparam: WPARAM) -> Option<LRESULT> {
     if !is_dark() {
         return None;
@@ -121,6 +130,9 @@ pub unsafe fn erasebkgnd(hwnd: HWND, wparam: WPARAM) -> Option<LRESULT> {
 /// gelap — notifikasi NM_CUSTOMDRAW header dikirim ke ListView, bukan ke window
 /// utama. Aman dipanggil berulang (uIdSubclass sama = perbarui, tak menumpuk).
 /// Proc menggating sendiri via `is_dark()`, jadi tetap benar saat tema terang.
+///
+/// # Safety
+/// `lv` harus handle ListView (`SysListView32`) valid.
 pub unsafe fn install_header(lv: HWND) {
     let _ = SetWindowSubclass(lv, Some(lv_subclass), HEADER_SUBCLASS_ID, 0);
     // Paksa header repaint agar warna langsung ikut.
@@ -182,7 +194,7 @@ unsafe fn header_customdraw(lparam: LPARAM) -> Option<LRESULT> {
         cchTextMax: 128,
         ..Default::default()
     };
-    let idx = cd.dwItemSpec as usize;
+    let idx = cd.dwItemSpec;
     SendMessageW(
         cd.hdr.hwndFrom,
         HDM_GETITEMW,
